+++ Source/Bumpy.Tests/CommandParserTests.cs
@@ -1,4 +1,5 @@
 ﻿using System.IO;
+using Bumpy.Config;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
 namespace Bumpy.Tests
@@ -92,7 +93,7 @@ namespace Bumpy.Tests
             var arguments = Parse("list");
 
             Assert.AreEqual(string.Empty, arguments.Profile);
-            Assert.AreEqual(new FileInfo("bumpy.yaml").FullName, arguments.ConfigFile.FullName);
+            Assert.AreEqual(new FileInfo(BumpyConfig.ConfigFile).FullName, arguments.ConfigFile.FullName);
             Assert.AreEqual(new DirectoryInfo(".").FullName, arguments.WorkingDirectory.FullName);
         }