+++ Source/Bumpy.Tests/CommandsTests.cs
@@ -170,7 +170,7 @@ namespace Bumpy.Tests
             fileUtil = fileUtil ?? Substitute.For<IFileUtil>();
             writeLine = writeLine ?? (l => { });
 
-            return new Commands(fileUtil, new FileInfo("bumpy.yaml"), new DirectoryInfo("."), noOperation, writeLine);
+            return new Commands(fileUtil, new FileInfo(BumpyConfig.ConfigFile), new DirectoryInfo("."), noOperation, writeLine);
         }
     }
 }