+++ Source/Bumpy/Config/BumpyConfig.cs
@@ -1,14 +1,11 @@
 ﻿using System.Collections.Generic;
-using YamlDotNet.Serialization;
 
 namespace Bumpy.Config
 {
     public class BumpyConfig
     {
-        public const string ConfigFile = "bumpy.yaml";
-        public const string LegacyConfigFile = ".bumpyconfig";
+        public const string ConfigFile = ".bumpyconfig";
 
-        [YamlMember(Alias = "queries")]
         public IEnumerable<BumpyConfigEntry> Queries { get; set; }
     }
 }