+++ CHANGELOG.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Changed
 
-- **Breaking Change:** The default configuration file to `bumpy.yaml`. The old format `.bumpyconfig` is still supported, but a warning is printed
+- **Breaking Change:** The configuration file format in `.bumpyconfig`. The old format is still supported, but a legacy warning is printed
 - Console output of all persisting commands (`increment`, `assign`, ...) to also display profiles
 - Console output of `ensure`. It will now print an overview of profiles with their version